conformance/buffimgs/cl_commandQueue_enqueueCopyImageToBuffer.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueCopyImageToBuffer conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<div style="position:absolute; left:0px; top:0px; visibility:hidden">
    <canvas id="dstBufferCanvas">
</div>
<script>
try {
    debug("WebCLCommandQueue :: enqueueCopyImageToBuffer() test case.");
    debug("( WebCLImage srcImage, WebCLBuffer dstBuffer, sequence<CLuint> srcOrigin, sequence<CLuint> srcRegion, CLuint dstOffset, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var origin = [0, 0];
    var region = [10, 10];
    var width = 10;
    var height = 10;

    var webCLImageDescriptorReadWrite = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, width, height);
    var webCLImageRead = webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite);
    var webCLImageReadWrite = webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite);
    var webCLImageWrite = webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite);

    // Testing for MEM_WRITE_ONLY Buffer.
    var webCLBufferWrite = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, Float32Array.BYTES_PER_ELEMENT * width * height * 4);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageRead, webCLBufferWrite, origin, region, 0);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferWrite, origin, region, 0);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageWrite, webCLBufferWrite, origin, region, Float32Array.BYTES_PER_ELEMENT * width * 4);");

    // Testing for MEM_READ_WRITE Buffer.
    var webCLBufferReadWrite = webCLContext.createBuffer(webcl.MEM_READ_WRITE, Float32Array.BYTES_PER_ELEMENT * width * height * 4);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageRead, webCLBufferReadWrite, origin, region, 0);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, region, 0);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageWrite, webCLBufferReadWrite, origin, region, Float32Array.BYTES_PER_ELEMENT * width * 4);");

    // Testing for MEM_READ_ONLY Buffer.
    var webCLBufferRead = webCLContext.createBuffer(webcl.MEM_READ_ONLY, Float32Array.BYTES_PER_ELEMENT * width * height * 4);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageRead, webCLBufferRead, origin, region, 0);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferRead, origin, region, 0);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageWrite, webCLBufferRead, origin, region, 0);");

    // Testing for Exceptions.
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(invalid_object, webCLBufferReadWrite, origin, region, 0);", "INVALID_MEM_OBJECT");
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, invalid_object, origin, region, 0);", "INVALID_MEM_OBJECT");

    // Testing for values of origin.
    var originInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, originInvalid, region, 0);", "INVALID_VALUE");

    // Testing for values of region.
    var regionInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLImageReadWrite, webCLBufferReadWrite, origin, regionInvalid, Float32Array.BYTES_PER_ELEMENT * width * 4);", "INVALID_VALUE");

    // Testing for values of destination Offset.
    var dstOffsetInvalid = width * height;
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImageToBuffer(webCLBufferReadWrite, webCLImageReadWrite, origin, region, dstOffsetInvalid);", "INVALID_VALUE");

    // FIXME :: Need to add test cases passing event parameters.

    // FIXME To add tet cases to verufy for exception INVALID_IMAGE_SIZE.

} catch (e) {
    testFailed("WebCLCommandQueue :: enqueueCopyImageToBuffer() throws an exception : " + e.message);
}
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
